% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingfunctions.R
\name{assert_non_empty_string}
\alias{assert_non_empty_string}
\title{Check object is a non-empty string}
\usage{
assert_non_empty_string(object)
}
\arguments{
\item{object}{Some value you want to assert is a non-empty string}
}
\value{
invisible(TRUE) if the object is a non-empty string. Throws an error if it is not
}
\description{
Check object is a non-empty string
}
\examples{
possiblestring = "Billy"
assert_non_empty_string(possiblestring)

}
\seealso{
Other customassertions: 
\code{\link{assert_that}()}
}
\concept{customassertions}
