% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfunctions.R
\name{checkScriptsDeclareTheirLocation}
\alias{checkScriptsDeclareTheirLocation}
\title{Finding Project Path, Check Directory Is Setup Correctly}
\usage{
checkScriptsDeclareTheirLocation(project_home_path)
}
\arguments{
\item{project_home_path}{path to project home (string)}
}
\value{
TRUE if project is ready for getpathProjectRootDirectory(), FALSE if some scripts need the line 'here::i_am(<path_from_project_root>)' to be added.
}
\description{
Before running \strong{getpathProjectRootDirectory()}, run \strong{checkScriptsDeclareTheirLocation(<project_dir>)} interactively.
This function searches <project_dir> recursively for R scripts and checks if the line 'here::i_am(<path_from_project_root>)' is present in ALL of them.
If not, a warning is output, recommending you add it line.
If all scripts declare their location relative to project root as above, \strong{getpathProjectRootDirectory()} is less likely to make a mistake.
}
\examples{
checkScriptsDeclareTheirLocation(
  system.file(
    "filesfortesting/here_positive_control", 
    package="utilitybelt"
  )
)
}
