% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classcheckingfunctions.R
\name{class_is_terminal_connection}
\alias{class_is_terminal_connection}
\title{class_is_terminal_connection}
\usage{
class_is_terminal_connection(object)
}
\arguments{
\item{object}{some object whose class you want to test}
}
\value{
TRUE/FALSE depending on whether the passed object is a connection to a terminal (logical)
}
\description{
Checks if an object is a connection to a terminal (e.g stderr() or stdout()).
}
\examples{
class_is_terminal_connection(stdout()) #True
class_is_terminal_connection(stderr()) #True
class_is_terminal_connection("random text") #false
}
