% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframefunctions.R
\name{df_sum_last_column}
\alias{df_sum_last_column}
\title{df_sum_last_column}
\usage{
df_sum_last_column(data)
}
\arguments{
\item{data}{a dataframe whose last column you want to sum (dataframe)}
}
\value{
sum of last column (numeric)
}
\description{
df_sum_last_column
}
\examples{
df_sum_last_column(mtcars)
}
