% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystemfunctions.R
\name{dir_create}
\alias{dir_create}
\title{Create directory
Take a path and create a new directory, with informative messages if it already exists.}
\usage{
dir_create(
  dir_path,
  force_overwrite = FALSE,
  promt_if_directory_exists = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dir_path}{path of new directory (string)}

\item{force_overwrite}{overwrite directory if it already exists (logical)}

\item{promt_if_directory_exists}{a message printed to stderr if directory already exists and force_overwrite = FALSE (string).  If writing a commandline script, for example, you might want to print "Use the -f flag to overwrite the existing directory".}

\item{verbose}{Print more info when running (logical)}
}
\value{
path to successfully created directory (string). If directory cannot be created, the function will throw an error.
}
\description{
Create directory
Take a path and create a new directory, with informative messages if it already exists.
}
