% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/favourites.R
\name{fav_shiny_renderDataTable_options}
\alias{fav_shiny_renderDataTable_options}
\title{Favorites}
\usage{
fav_shiny_renderDataTable_options()
}
\value{
Example command using favorite renderDataTable options (string)
}
\description{
List favorite fav_shiny_renderDataTable_options
}
\examples{
fav_shiny_renderDataTable_options()
}
