% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystemfunctions.R
\name{file_create}
\alias{file_create}
\title{Create file}
\usage{
file_create(
  file_path,
  force_overwrite = FALSE,
  promt_if_file_exists = NULL,
  verbose = TRUE
)
}
\arguments{
\item{file_path}{path of new file (string)}

\item{force_overwrite}{overwrite file if it already exists (logical)}

\item{promt_if_file_exists}{a message printed to stderr if file already exists and force_overwrite = FALSE  (in addition to some generic error message).  If writing a commandline script, for example, you might want to print "Use the -f flag to overwrite the existing file". (string)}

\item{verbose}{Print more info when running (logical)}
}
\value{
path to successfully created file (string). If file cannot be created, the function will throw an error.
}
\description{
Take a path and create a new file, with informative messages if it already exists.
}
\seealso{
Other filesystemfunctions: 
\code{\link{dir_create}()}
}
\concept{filesystemfunctions}
