% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystemfunctions.R
\name{files_exist_all}
\alias{files_exist_all}
\title{Check files ALL exist}
\usage{
files_exist_all(files)
}
\arguments{
\item{files}{a vector of filenames/filepaths. Directories and Files are treated the exact same ways (character).}
}
\value{
do all files/directories in \strong{files} vector exist: TRUE/FALSE (logical)
}
\description{
Check files ALL exist
}
