% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filesystemfunctions.R
\name{files_exist_which}
\alias{files_exist_which}
\title{Check which files exist}
\usage{
files_exist_which(files)
}
\arguments{
\item{files}{a vector of filenames/filepaths. Directories and Files are treated the exact same ways (character).}
}
\value{
vector indicating which files/directories in \strong{files} exist: TRUE/FALSE (logical)
}
\description{
Check which files exist
}
