% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfunctions.R
\name{getpathProjectRootDirectory}
\alias{getpathProjectRootDirectory}
\title{getpathProjectRootDirectory}
\usage{
getpathProjectRootDirectory(warn = TRUE)
}
\arguments{
\item{warn}{Warn user to run checkScriptsDeclareTheirLocation(<project_directory>) to ensure its set up to accurately identify a projects 'home'  (logical)}
}
\value{
absolute path of the project directory rootfolder
}
\description{
getpathProjectRootDirectory
}
\examples{
getpathProjectRootDirectory(warn=FALSE)
}
