% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfunctions.R
\name{getpathSourceDirectory}
\alias{getpathSourceDirectory}
\title{Get source directory (ONLY WORKS IF FILE IS SOURCED!)}
\usage{
getpathSourceDirectory(warn = FALSE)
}
\arguments{
\item{warn}{warn the user that running interactively will error? (lgl)}
}
\value{
path to the directory containing the script from which the function is called. If run interactively, will throw an error (string)
}
\description{
WARNING: if you run in interactive mode, this function will print an informative error
}
