% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/librarymanagementfunctions.R
\name{library_quietload}
\alias{library_quietload}
\title{Loads a library quietly
This function loads a package, optionally outputting a single line message if successful.
Will always print the full error message if load fails}
\usage{
library_quietload(library_name, verbose = FALSE)
}
\arguments{
\item{library_name}{(string) Name of library to load}

\item{verbose}{(logical) Should we print a message if load was successful?}
}
\value{
void. This function is run for its side effect.
}
\description{
Loads a library quietly
This function loads a package, optionally outputting a single line message if successful.
Will always print the full error message if load fails
}
\examples{
"library_quietload(dplyr, verbose=TRUE)"
}
