% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfunctions.R
\name{path_process_vectorized}
\alias{path_process_vectorized}
\title{Process filenames/paths}
\usage{
path_process_vectorized(
  paths,
  extract_basename,
  remove_extension,
  verbose = FALSE
)
}
\arguments{
\item{paths}{a vector of filenames/filepaths (character)}

\item{extract_basename}{extract basename (logical)}

\item{remove_extension}{remove extension (logical)}

\item{verbose}{(logical)}
}
\value{
vector of processed filenames (character)
}
\description{
Takes a filename and removes extensions (file.vep.tsv to file.vep and file.vep.vcf.gz to file.vep AND/OR  extracts basename (i.e. dir1/dir2/file.tsv to file)
}
\examples{
path_process_vectorized(
paths = c("Dir1/Dir2/Dir3/file.tsv", "/root/dir/myfile"), 
extract_basename = TRUE, 
remove_extension = TRUE)
}
