% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optparseextensions.R
\name{print_help_stderr}
\alias{print_help_stderr}
\title{Print Help to Stderr}
\usage{
print_help_stderr()
}
\value{
invisible (NULL)
}
\description{
Wraps around optparse's print_help function but sends help info to stderr
}
