% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggextensions.R
\name{theme_common_adjustments}
\alias{theme_common_adjustments}
\title{Custom Themes}
\usage{
theme_common_adjustments(
  dist_from_plot_xlab = 10,
  dist_from_plot_ylab = 10,
  dist_from_plot_ggtitle = 10,
  no_background = FALSE
)
}
\arguments{
\item{dist_from_plot_xlab}{distance between x axis title and plot (number)}

\item{dist_from_plot_ylab}{distance between y axis title and plot (number)}

\item{dist_from_plot_ggtitle}{distance between y axis title and plot (number)}

\item{no_background}{set all backgrounds to clear? (flag)}
}
\value{
ggtheme
}
\description{
ggplot theme that fixes a bunch of issues I tend to have with plots
}
\examples{
mtcars \%>\% 
  ggplot2::ggplot(ggplot2::aes(cyl>6, mpg)) + 
  ggplot2::geom_point() + 
  theme_axis_titles()
}
