% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfunctions.R
\name{thisFileWasSourced}
\alias{thisFileWasSourced}
\title{Was this file sourced or run interactively?}
\usage{
thisFileWasSourced()
}
\value{
TRUE if script from which file was called was sourced, FALSE if run interactively (e.g. line by line in rstudio) (logical)
}
\description{
Was this file sourced or run interactively?
}
\examples{
thisFileWasSourced()
}
