% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathfunctions.R
\name{path_process}
\alias{path_process}
\title{Process filenames/paths}
\usage{
path_process(path, extract_basename, remove_extension, verbose = FALSE)
}
\arguments{
\item{path}{a filename/filepath (string)}

\item{extract_basename}{extract basename (logical)}

\item{remove_extension}{remove extension (logical)}

\item{verbose}{(logical)}
}
\value{
processed path (string)
}
\description{
Process filenames/paths
}
\examples{
path_process("Dir1/Dir2/Dir3/file.tsv", extract_basename=TRUE, remove_extension=TRUE)
}
