% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggextensions.R
\name{theme_axis_titles}
\alias{theme_axis_titles}
\title{Custom Themes}
\usage{
theme_axis_titles(
  dist_from_plot_xlab = 10,
  dist_from_plot_ylab = 10,
  dist_from_plot_ggtitle = 10
)
}
\value{
ggtheme
}
\description{
Custom Themes
}
